Create LongestSubstrWithUniqueChars2.java
This is essentially the same program as "LongestSubstrWithUniqueChars2.java", but the problem with previous problem was that java by default only provides hashmap which is a javas alternative to cpp map, but the problem of cpp map is that it can't have different values with same key the same problem persists with hashmap. So the previous solution can only show one substring of longest length. So to solve that problem we merely switched keys with values, becuz key can't be duplicate but values can be.